Push

+++ M2/FdF 🎯️/C/Utils.c
@@ -18,7 +18,7 @@ int	put_pixel(t_data *data, t_point p)
 
     if (p.x >= 0 && p.x < W_WIDTH && p.y >= 0 && p.y < W_HEIGHT)
     {
-        dst = data->addr + (p.y * data->line_length + p.x * (data->bits_per_pixel / 8));
+        dst = data->addr + (int)(p.y * data->line_length + p.x * (data->bits_per_pixel / 8));
         *(unsigned int *)dst = rgbtoi(p.color);
         return (1);
     }
@@ -54,12 +54,12 @@ void	link_points(t_data *img, t_tab tab)
 		{
 			if (x + 1 < tab.width)
 			{
-				printf("Drawing line from (%d, %d) to (%d, %d)\n", tab.tab[x][y].x, tab.tab[x][y].y, tab.tab[x + 1][y].x, tab.tab[x + 1][y].y);
+				printf("Drawing line from (%f, %f) to (%f, %f)\n", tab.tab[x][y].x, tab.tab[x][y].y, tab.tab[x + 1][y].x, tab.tab[x + 1][y].y);
 				draw_line(img, tab.tab[x][y], tab.tab[x + 1][y]);
 			}
 			if (y + 1 < tab.height)
 			{
-				printf("Drawing line from (%d, %d) to (%d, %d)\n", tab.tab[x][y].x, tab.tab[x][y].y, tab.tab[x][y + 1].x, tab.tab[x][y + 1].y);
+				printf("Drawing line from (%f, %f) to (%f, %f)\n", tab.tab[x][y].x, tab.tab[x][y].y, tab.tab[x][y + 1].x, tab.tab[x][y + 1].y);
 				draw_line(img, tab.tab[x][y], tab.tab[x][y + 1]);
 			}
 			x++;
